Crate html5ever

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Reexports

pub use markup5ever::*;

pub use driver::ParseOpts;

pub use driver::parse_document;

pub use driver::parse_fragment;

pub use driver::Parser;

pub use serialize::serialize;

Modules

driver 	 [
Experimental
]
High-level interface to the parser.

serialize 	 [
Experimental
]

tokenizer 	 [
Experimental
]
The HTML5 tokenizer.

tree_builder 	 [
Experimental
]
The HTML5 tree builder.

Macros

expanded_name

local_name

namespace_prefix

namespace_url

ns

small_char_set