Module html5ever::buffer_queue

The BufferQueue struct and helper types.

This type is designed for the efficient parsing of string data, especially where many significant characters are from the ascii range 0-63. This includes, for example, important characters in xml/html parsing.

Good and predictable performance is achieved by avoiding allocation where possible (a.k.a. zero copy).

Structs

BufferQueue

A queue of owned string buffers, which supports incrementally consuming characters.

Enums

SetResult

Result from pop_except_from containing either a character from a SmallCharSet, or a string buffer of characters not from the set.