Struct html5ever::driver::Parser

pub struct Parser<Sink> where
    Sink: TreeSink,  {
    pub tokenizer: Tokenizer<TreeBuilder<Sink::Handle, Sink>>,
    pub input_buffer: BufferQueue,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

An HTML parser, ready to receive Unicode input through the tendril::TendrilSink trait’s methods.

Fields
tokenizer: Tokenizer<TreeBuilder<Sink::Handle, Sink>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

input_buffer: BufferQueue

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<Sink: TreeSink> Parser<Sink>

fn from_utf8(self) -> Utf8LossyDecoder<Self>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Wrap this parser into a TendrilSink that accepts UTF-8 bytes.

Use this when your input is bytes that are known to be in the UTF-8 encoding. Decoding is lossy, like String::from_utf8_lossy.