Module html5ever::encoding

Re-export the rust-encoding crate.

Modules

all
A list of all supported encodings. Useful for encodings fixed in the compile time.

codec
Codec implementations.

label
An interface for retrieving an encoding (or a set of encodings) from a string/numeric label.

types
Interface to the character encoding.

Structs

CodecError
Error information from either encoder or decoder.

Enums

DecoderTrap
Trap, which handles decoder errors.

EncoderTrap

Traits

ByteWriter
Byte writer used by encoders. In most cases this will be an owned vector of u8.

Encoding
Character encoding.

RawDecoder
Decoder converting a byte sequence into a Unicode string. This is a lower level interface, and normally Encoding::decode should be used instead.

RawEncoder
Encoder converting a Unicode string into a byte sequence. This is a lower level interface, and normally Encoding::encode should be used instead.

StringWriter
String writer used by decoders. In most cases this will be an owned string.

Functions

decode
Determine the encoding by looking for a Byte Order Mark (BOM) and decoded a single string in memory. Return the result and the used encoding.

Type Definitions

DecoderTrapFunc

EncoderTrapFunc

EncodingRef