html5ever

Module interface

Types for tag and attribute names, and tree-builder functionality.

Modules

tree_builder

This module contains functionality for managing the DOM, including adding/removing nodes.

Structs

Attribute

A tag attribute, e.g. class="test" in <div class="test" ...>.

ElementFlags

Special properties of an element, useful for tagging elements with this information.

ExpandedName

An expanded name, containing the tag and the namespace.

QualName

A fully qualified name (with a namespace), used to depict names of tags and attributes.

Enums

NodeOrText

Something which can be inserted into the DOM.

QuirksMode

A document’s quirks mode, for compatibility with old browsers. See quirks mode on wikipedia for more information.

TokenizerResult

Traits

ElemName

An abstraction over any type that can represent an element’s local name and namespace.

Tracer

Trace hooks for a garbage-collected DOM.

TreeSink

Methods a parser can use to create the DOM. The DOM provider implements this trait.

Functions

create_element

A constructor for an element.