Trait html5ever::serialize::Serialize

pub trait Serialize {
    fn serialize<S>(
        &self, 
        serializer: &mut S, 
        traversal_scope: TraversalScope
    ) -> Result<(), Error>
    where
        S: Serializer;
}

Types that can be serialized (according to the xml-like scheme in Serializer) implement this trait.

Required methods

fn serialize<S>(
&self,
serializer: &mut S,
traversal_scope: TraversalScope
) -> Result<(), Error> where
S: Serializer,

Take the serializer and call its methods to serialize this type. The type will dictate which methods are called and with what parameters.