Trait html5ever::serialize::Serializer

pub trait Serializer {
    fn start_elem<'a, AttrIter>(
        &mut self, 
        name: QualName, 
        attrs: AttrIter
    ) -> Result<(), Error>
    where
        AttrIter: Iterator<Item = (&'a QualName, &'a str)>;

    fn end_elem(&mut self, name: QualName) -> Result<(), Error>;

    fn write_text(&mut self, text: &str) -> Result<(), Error>;

    fn write_comment(&mut self, text: &str) -> Result<(), Error>;

    fn write_doctype(&mut self, name: &str) -> Result<(), Error>;

    fn write_processing_instruction(
        &mut self, 
        target: &str, 
        data: &str
    ) -> Result<(), Error>;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn start_elem<'a, AttrIter>(
&mut self,
name: QualName,
attrs: AttrIter
) -> Result<(), Error> where
AttrIter: Iterator<Item = (&'a QualName, &'a str)>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn end_elem(&mut self, name: QualName) -> Result<(), Error>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn write_text(&mut self, text: &str) -> Result<(), Error>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn write_comment(&mut self, text: &str) -> Result<(), Error>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn write_doctype(&mut self, name: &str) -> Result<(), Error>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn write_processing_instruction(
&mut self,
target: &str,
data: &str
) -> Result<(), Error>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?