Module html5ever::sink::owned_dom

A simple DOM where every node is owned by its parent.

Since ownership is more complicated during parsing, we actually build a different type and then transmute to the public Node. This is believed to be memory safe, but if you want to be extra careful you can use RcDom instead.

Structs

Handle

Node

OwnedDom

Sink

SquishyNode
The internal type we use for nodes during parsing.