Struct html5ever::tendril::stream::Utf8LossyDecoder

pub struct Utf8LossyDecoder<Sink, A = NonAtomic> where
    A: Atomicity,
    Sink: TendrilSink<UTF8, A>,  {
    pub inner_sink: Sink,
    // some fields omitted
}

A TendrilSink adaptor that takes bytes, decodes them as UTF-8, lossily replace ill-formed byte sequences with U+FFFD replacement characters, and emits Unicode (StrTendril).

This does not allocate memory: the output is either subtendrils on the input, on inline tendrils for a single code point.

Fields
inner_sink: Sink
Methods

impl<Sink, A> Utf8LossyDecoder<Sink, A> where
A: Atomicity,
Sink: TendrilSink<UTF8, A>,

pub fn new(inner_sink: Sink) -> Utf8LossyDecoder<Sink, A>

Create a new incremental UTF-8 decoder.