Enum html5ever::tokenizer::SetResult

pub enum SetResult {
    FromSet(char),
    NotFromSet(Tendril<UTF8, NonAtomic>),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Result from pop_except_from.

Variants
FromSet(char)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

NotFromSet(Tendril<UTF8, NonAtomic>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?