html5ever::tokenizer::interface

Trait TokenSink

pub trait TokenSink {
    type Handle;

    // Required method
    fn process_token(
        &self,
        token: Token,
        line_number: u64,
    ) -> TokenSinkResult<Self::Handle>;

    // Provided methods
    fn end(&self) { ... }
    fn adjusted_current_node_present_but_not_in_html_namespace(&self) -> bool { ... }
}

Types which can receive tokens from the tokenizer.

Required Associated Types

type Handle

The type of a DOM node.

Required Methods

fn process_token( &self, token: Token, line_number: u64, ) -> TokenSinkResult<Self::Handle>

Process a token.

Provided Methods

fn end(&self)

Signal that tokenization reached the end of the document.

fn adjusted_current_node_present_but_not_in_html_namespace(&self) -> bool

Used in the markup declaration open state. By default, this always returns false and thus all CDATA sections are tokenized as bogus comments.