Struct html5ever::tokenizer::BufferQueue

pub struct BufferQueue { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A queue of owned string buffers, which supports incrementally consuming characters.

Methods

impl BufferQueue

fn new() -> BufferQueue

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Create an empty BufferQueue.

fn is_empty(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns whether the queue is empty.

fn pop_front(&mut self) -> Option<Tendril<UTF8, NonAtomic>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Get the tendril at the beginning of the queue.

fn push_front(&mut self, buf: Tendril<UTF8, NonAtomic>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Add a buffer to the beginning of the queue.

fn push_back(&mut self, buf: Tendril<UTF8, NonAtomic>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Add a buffer to the end of the queue.

fn peek(&self) -> Option<char>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Look at the next available character, if any.

fn next(&mut self) -> Option<char>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Get the next character, if one is available.

fn pop_except_from(&mut self, set: SmallCharSet) -> Option<SetResult>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Pops and returns either a single character from the given set, or a StrTendril of characters none of which are in the set. The set is represented as a bitmask and so can only contain the first 64 ASCII characters.

fn eat<F>(&mut self, pat: &str, eq: F) -> Option<bool> where
F: Fn(&u8, &u8) -> bool,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?