Struct html5ever::tokenizer::TokenizerOpts

pub struct TokenizerOpts {
    pub exact_errors: bool,
    pub discard_bom: bool,
    pub profile: bool,
    pub initial_state: Option<State>,
    pub last_start_tag_name: Option<String>,
}

Tokenizer options, with an impl for Default.

Fields
exact_errors: bool

Report all parse errors described in the spec, at some performance penalty? Default: false

discard_bom: bool

Discard a U+FEFF BYTE ORDER MARK if we see one at the beginning of the stream? Default: true

profile: bool

Keep a record of how long we spent in each state? Printed when end() is called. Default: false

initial_state: Option<State>

Initial state override. Only the test runner should use a non-None value!

last_start_tag_name: Option<String>

Last start tag. Only the test runner should use a non-None value!

FIXME: Can't use Tendril because we want TokenizerOpts to be Send.