Trait html5ever::tokenizer::TokenSink

pub trait TokenSink {
    type Handle;
    fn process_token(
        &mut self, 
        token: Token, 
        line_number: u64
    ) -> TokenSinkResult<Self::Handle>;

    fn end(&mut self) { ... }

    fn adjusted_current_node_present_but_not_in_html_namespace(&self) -> bool { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Types which can receive tokens from the tokenizer.

Associated Types

type Handle

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn process_token(
&mut self,
token: Token,
line_number: u64
) -> TokenSinkResult<Self::Handle>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Process a token.

Provided Methods

fn end(&mut self)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn adjusted_current_node_present_but_not_in_html_namespace(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Used in the markup declaration open state. By default, this always returns false and thus all CDATA sections are tokenized as bogus comments. https://html.spec.whatwg.org/multipage/#markup-declaration-open-state