Enum html5ever::tree_builder::interface::NodeOrText

pub enum NodeOrText<Handle> {
    AppendNode(Handle),
    AppendText(String),
}

Something which can be inserted into the DOM.

Adjacent sibling text nodes are merged into a single node, so the sink may not want to allocate a Handle for each.

Variants

AppendNode

AppendText