Struct html5ever::tree_builder::ElementFlags

pub struct ElementFlags {
    pub template: bool,
    pub mathml_annotation_xml_integration_point: bool,
    // some fields omitted
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
template: bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A document fragment should be created, associated with the element, and returned in TreeSink::get_template_contents

https://html.spec.whatwg.org/multipage/#template-contents

mathml_annotation_xml_integration_point: bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

This boolean should be recorded with the element and returned in TreeSink::is_mathml_annotation_xml_integration_point

https://html.spec.whatwg.org/multipage/#html-integration-point