Struct html5ever::tree_builder::TreeBuilderOpts

pub struct TreeBuilderOpts {
    pub exact_errors: bool,
    pub scripting_enabled: bool,
    pub iframe_srcdoc: bool,
    pub drop_doctype: bool,
    pub ignore_missing_rules: bool,
    pub quirks_mode: QuirksMode,
}

Tree builder options, with an impl for Default.

Fields

exact_errors

Report all parse errors described in the spec, at some performance penalty? Default: false

scripting_enabled

Is scripting enabled?

iframe_srcdoc

Is this an iframe srcdoc document?

drop_doctype

Should we drop the DOCTYPE (if any) from the tree?

ignore_missing_rules

The <svg>, <math>, and <template> tags have special parsing rules that are currently unimplemented. By default we panic!() if any of these tags is encountered. If this option is enabled, we will instead attempt to parse them using the ordinary HTML parsing rules.

Warning: This may produce extremely incorrect results on some documents!

quirks_mode

Initial TreeBuilder quirks mode. Default: NoQuirks