Crate html5ever

Re-exports

pub use driver::parse_document;

pub use driver::parse_fragment;

pub use driver::ParseOpts;

pub use driver::Parser;

pub use serialize::serialize;

Modules

buffer_queue
The BufferQueue struct and helper types.

data
Data that is known at compile-time and hard-coded into the binary.

driver
High-level interface to the parser.

interface
Types for tag and attribute names, and tree-builder functionality.

rcdom
A simple reference-counted DOM.

serialize
Traits for serializing elements. The serializer expects the data to be xml-like (with a name, and optional children, attrs, text, comments, doctypes, and processing instructions). It uses the visitor pattern, where the serializer and the serializable objects are decoupled and implement their own traits.

serialize

smallcharset
This module contains a single struct SmallCharSet. See its documentation for details.

tendril

tokenizer
The HTML5 tokenizer.

tree_builder
The HTML5 tree builder.

Macros

expanded_name
Helper to quickly create an expanded name.

local_name
Takes a local name as a string and returns its key in the string cache.

namespace_prefix
Takes a namespace prefix string and returns its key in a string cache.

namespace_url
Takes a namespace url string and returns its key in a string cache.

ns
Maps the input of namespace_prefix! to the output of namespace_url!.

small_char_set
Create a SmallCharSet, with each space-separated number stored in the set.

Structs

LocalNameStaticSet

NamespaceStaticSet

PrefixStaticSet

SmallCharSet
Represents a set of "small characters", those with Unicode scalar values less than 64.

Type Definitions

LocalName

Namespace

Prefix