Module html5ever::driver

High-level interface to the parser.

Structs

ParseOpts
All-encompassing options struct for the parser.

Parser
An HTML parser, ready to receive Unicode input through the tendril::TendrilSink trait’s methods.

Functions

parse_document
Parse an HTML document

parse_fragment
Parse an HTML fragment

parse_fragment_for_element
Like parse_fragment, but with an existing context element and optionally a form element.