Function html5ever::driver::parse_document

pub fn parse_document<Sink>(sink: Sink, opts: ParseOpts) -> Parser<Sink>
where
    Sink: TreeSink,

Parse an HTML document

The returned value implements tendril::TendrilSink so that Unicode input may be provided incrementally, or all at once with the one method.

If your input is bytes, use Parser::from_utf8.