Function html5ever::driver::parse_fragment

pub fn parse_fragment<Sink>(
    sink: Sink, 
    opts: ParseOpts, 
    context_name: QualName, 
    context_attrs: Vec<Attribute>
) -> Parser<Sink> where
    Sink: TreeSink, 

Parse an HTML fragment

The returned value implements tendril::TendrilSink so that Unicode input may be provided incrementally, or all at once with the one method.

If your input is bytes, use Parser::from_utf8.