Struct html5ever::driver::BytesParser

pub struct BytesParser<Sink> where Sink: TreeSink {
    // some fields omitted
}

An HTML parser, ready to recieve bytes input through the tendril::TendrilSink trait’s methods.

See Parser::from_bytes.

Methods

impl<Sink: TreeSink> BytesParser<Sink>

fn str_parser(&self) -> &Parser<Sink>

Access the underlying Parser

fn str_parser_mut(&mut self) -> &mut Parser<Sink>

Access the underlying Parser

fn process_unicode(&mut self, t: StrTendril)

Insert a Unicode chunk in the middle of the byte stream.

This is e.g. for supporting document.write.