Struct html5ever::driver::Parser

pub struct Parser<Sink> where Sink: TreeSink {
    pub tokenizer: Tokenizer<TreeBuilder<Sink::Handle, Sink>>,
}

An HTML parser, ready to recieve Unicode input through the tendril::TendrilSink trait’s methods.

Fields
tokenizer: Tokenizer<TreeBuilder<Sink::Handle, Sink>>
Methods

impl<Sink: TreeSink> Parser<Sink>

fn from_utf8(self) -> Utf8LossyDecoder<Self>

Wrap this parser into a TendrilSink that accepts UTF-8 bytes.

Use this when your input is bytes that are known to be in the UTF-8 encoding. Decoding is lossy, like String::from_utf8_lossy.

fn from_bytes(self, opts: BytesOpts) -> BytesParser<Sink>

Wrap this parser into a TendrilSink that accepts bytes and tries to detect the correct character encoding.

Currently this looks for a Byte Order Mark, then uses BytesOpts::transport_layer_encoding, then falls back to UTF-8.

FIXME(https://github.com/servo/html5ever/issues/18): this should look for <meta> elements and other data per https://html.spec.whatwg.org/multipage/syntax.html#determining-the-character-encoding