Struct html5ever::driver::Parser

pub struct Parser<Sink>
where
    Sink: TreeSink,
{
    pub tokenizer: Tokenizer<TreeBuilder<Sink::Handle, Sink>>,
    pub input_buffer: BufferQueue,
}

An HTML parser, ready to receive Unicode input through the tendril::TendrilSink trait’s methods.

Fields
tokenizer: Tokenizer<TreeBuilder<Sink::Handle, Sink>>input_buffer: BufferQueue
Implementations

impl<Sink: TreeSink> Parser<Sink>

pub fn from_utf8(self) -> Utf8LossyDecoder<Self>

Wrap this parser into a TendrilSink that accepts UTF-8 bytes.

Use this when your input is bytes that are known to be in the UTF-8 encoding. Decoding is lossy, like String::from_utf8_lossy.