Module html5ever::serialize

Structs

SerializeOpts

Enums

TraversalScope
Used as a parameter to serialize, telling it if we want to skip the parent.

Traits

Serialize
Types that can be serialized (according to the xml-like scheme in Serializer) implement this trait.

Serializer
Types that are capable of serializing implement this trait

Functions

serialize

Type Definitions

AttrRef
A type alias for an attribute name and value (e.g. the class="test" in <div class="test"> is represented as (<QualName of type class>, "test").