html5ever::serialize

Trait Serializer

pub trait Serializer {
    // Required methods
    fn start_elem<'a, AttrIter>(
        &mut self,
        name: QualName,
        attrs: AttrIter,
    ) -> Result<(), Error>
       where AttrIter: Iterator<Item = (&'a QualName, &'a str)>;
    fn end_elem(&mut self, name: QualName) -> Result<(), Error>;
    fn write_text(&mut self, text: &str) -> Result<(), Error>;
    fn write_comment(&mut self, text: &str) -> Result<(), Error>;
    fn write_doctype(&mut self, name: &str) -> Result<(), Error>;
    fn write_processing_instruction(
        &mut self,
        target: &str,
        data: &str,
    ) -> Result<(), Error>;
}

Types that are capable of serializing implement this trait

Required Methods

fn start_elem<'a, AttrIter>( &mut self, name: QualName, attrs: AttrIter, ) -> Result<(), Error>
where AttrIter: Iterator<Item = (&'a QualName, &'a str)>,

Serialize the start of an element, for example <div class="test">.

fn end_elem(&mut self, name: QualName) -> Result<(), Error>

Serialize the end of an element, for example </div>.

fn write_text(&mut self, text: &str) -> Result<(), Error>

Serialize a plain text node.

fn write_comment(&mut self, text: &str) -> Result<(), Error>

Serialize a comment node, for example <!-- comment -->.

fn write_doctype(&mut self, name: &str) -> Result<(), Error>

Serialize a doctype node, for example <!doctype html>.

fn write_processing_instruction( &mut self, target: &str, data: &str, ) -> Result<(), Error>

Serialize a processing instruction node, for example <?xml-stylesheet type="text/xsl" href="style.xsl"?>.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.