html5ever

Module tokenizer

The HTML5 tokenizer.

Re-exports

pub use self::interface::CharacterTokens;

pub use self::interface::EOFToken;

pub use self::interface::NullCharacterToken;

pub use self::interface::ParseError;

pub use self::interface::CommentToken;

pub use self::interface::DoctypeToken;

pub use self::interface::TagToken;

pub use self::interface::EndTag;

pub use self::interface::StartTag;

pub use crate::buffer_queue::FromSet;

pub use crate::buffer_queue::NotFromSet;

Modules

states

Tokenizer states.

Structs

BufferQueue

A queue of owned string buffers, which supports incrementally consuming characters.

Doctype

A DOCTYPE token.

Tag

A tag token.

Tokenizer

The HTML tokenizer.

TokenizerOpts

Tokenizer options, with an impl for Default.

Enums

ProcessResult

SetResult

Result from pop_except_from containing either a character from a SmallCharSet, or a string buffer of characters not from the set.

TagKind

Token

TokenSinkResult

TokenizerResult

Traits

TokenSink

Types which can receive tokens from the tokenizer.