Module html5ever::tokenizer::states

Tokenizer states.

This is public for use by the tokenizer tests. Other library users should not have to care about this.

Re-exports

pub use self::AttrValueKind::*;

pub use self::DoctypeIdKind::*;

pub use self::RawKind::*;

pub use self::ScriptEscapeKind::*;

pub use self::State::*;

Enums

AttrValueKind

DoctypeIdKind

RawKind

ScriptEscapeKind

State