Struct html5ever::tokenizer::Tokenizer

pub struct Tokenizer<Sink> {
    // some fields omitted
}

The HTML tokenizer.

Methods

impl<Sink: TokenSink> Tokenizer<Sink>

fn new(sink: Sink, opts: TokenizerOpts) -> Tokenizer<Sink>

Create a new tokenizer which feeds tokens to a particular TokenSink.

fn unwrap(self) -> Sink

fn sink<'a>(&'a self) -> &'a Sink

fn sink_mut<'a>(&'a mut self) -> &'a mut Sink

fn feed(&mut self, input: StrTendril)

Feed an input string into the tokenizer.

fn set_plaintext_state(&mut self)

fn run(&mut self)

Run the state machine for as long as we can.

impl<Sink: TokenSink> Tokenizer<Sink>

fn end(&mut self)

Indicate that we have reached the end of the input.