Struct html5ever::tokenizer::Tokenizer

pub struct Tokenizer<Sink> {
    pub sink: Sink,
    // some fields omitted
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The HTML tokenizer.

Fields
sink: Sink

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Destination for tokens we emit.

Methods

impl<Sink: TokenSink> Tokenizer<Sink>

fn new(sink: Sink, opts: TokenizerOpts) -> Tokenizer<Sink>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Create a new tokenizer which feeds tokens to a particular TokenSink.

fn feed(&mut self, input: &mut BufferQueue) -> TokenizerResult<Sink::Handle>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Feed an input string into the tokenizer.

fn set_plaintext_state(&mut self)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl<Sink: TokenSink> Tokenizer<Sink>

fn end(&mut self)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Indicate that we have reached the end of the input.