Trait html5ever::tokenizer::TokenSink

pub trait TokenSink {
    fn process_token(&mut self, token: Token);

    fn query_state_change(&mut self) -> Option<State> { ... }
}

Types which can receive tokens from the tokenizer.

Required Methods

fn process_token(&mut self, token: Token)

Process a token.

Provided Methods

fn query_state_change(&mut self) -> Option<State>

The tokenizer will call this after emitting any start tag. This allows the tree builder to change the tokenizer's state. By default no state changes occur.