Module html5ever::tree_builder

The HTML5 tree builder.

Re-exports

pub use crate::interface::AppendNode;

pub use crate::interface::AppendText;

pub use crate::interface::LimitedQuirks;

pub use crate::interface::NoQuirks;

pub use crate::interface::Quirks;

pub use self::PushFlag::*;

Modules

data

tag_sets

Various sets of HTML tag names, and macros for declaring them.

types

Types used within the tree builder code. Not exported to users.

Structs

ActiveFormattingIter

Attribute

A tag attribute, e.g. class="test" in <div class="test" ...>.

ElementFlags

Special properties of an element, useful for tagging elements with this information.

TreeBuilder

The HTML tree builder.

TreeBuilderOpts

Tree builder options, with an impl for Default.

Enums

Bookmark

NextParserState

Whether to interrupt further parsing of the current input until the next explicit resumption of the tokenizer, or continue without any interruption.

NodeOrText

Something which can be inserted into the DOM.

PushFlag

QuirksMode

A document’s quirks mode, for compatibility with old browsers. See quirks mode on wikipedia for more information.

Traits

Tracer

Trace hooks for a garbage-collected DOM.

TreeSink

Methods a parser can use to create the DOM. The DOM provider implements this trait.

Functions

any_not_whitespace

create_element

A constructor for an element.

current_node

html_elem