Struct html5ever::tree_builder::TreeBuilder

pub struct TreeBuilder<Handle, Sink> {
    // some fields omitted
}

The HTML tree builder.

Methods

impl<Handle, Sink> TreeBuilder<Handle, Sink> where Handle: Clone, Sink: TreeSink<Handle=Handle>

fn new(sink: Sink, opts: TreeBuilderOpts) -> TreeBuilder<Handle, Sink>

Create a new tree builder which sends tree modifications to a particular TreeSink.

The tree builder is also a TokenSink.

fn new_for_fragment(sink: Sink, context_elem: Handle, form_elem: Option<Handle>, opts: TreeBuilderOpts) -> TreeBuilder<Handle, Sink>

Create a new tree builder which sends tree modifications to a particular TreeSink. This is for parsing fragments.

The tree builder is also a TokenSink.

fn tokenizer_state_for_context_elem(&self) -> State

fn unwrap(self) -> Sink

fn sink<'a>(&'a self) -> &'a Sink

fn sink_mut<'a>(&'a mut self) -> &'a mut Sink

fn trace_handles(&self, tracer: &Tracer<Handle=Handle>)

Call the Tracer's trace_handle method on every Handle in the tree builder's internal state. This is intended to support garbage-collected DOMs.

fn is_fragment(&self) -> bool

Are we parsing a HTML fragment?