Struct html5ever::tree_builder::TreeBuilder

pub struct TreeBuilder<Handle, Sink> {
    pub sink: Sink,
    // some fields omitted
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The HTML tree builder.

Fields
sink: Sink

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Consumer of tree modifications.

Methods

impl<Handle, Sink> TreeBuilder<Handle, Sink> where
Handle: Clone,
Sink: TreeSink<Handle = Handle>,

fn new(sink: Sink, opts: TreeBuilderOpts) -> TreeBuilder<Handle, Sink>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Create a new tree builder which sends tree modifications to a particular TreeSink.

The tree builder is also a TokenSink.

fn new_for_fragment(
sink: Sink,
context_elem: Handle,
form_elem: Option<Handle>,
opts: TreeBuilderOpts
) -> TreeBuilder<Handle, Sink>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Create a new tree builder which sends tree modifications to a particular TreeSink. This is for parsing fragments.

The tree builder is also a TokenSink.

fn tokenizer_state_for_context_elem(&self) -> State

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn trace_handles(&self, tracer: &Tracer<Handle = Handle>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Call the Tracer's trace_handle method on every Handle in the tree builder's internal state. This is intended to support garbage-collected DOMs.

fn is_fragment(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Are we parsing a HTML fragment?